Export binding-specific config types
