+++ incubator/exporter-otlp/src/main/java/io/aklivity/zilla/runtime/exporter/otlp/config/OtlpOverridesConfig.java
@@ -12,30 +12,19 @@
  * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package io.aklivity.zilla.runtime.exporter.otlp.internal.config;
+package io.aklivity.zilla.runtime.exporter.otlp.config;
 
-import java.util.Set;
+import java.net.URI;
 
 import io.aklivity.zilla.runtime.engine.config.OptionsConfig;
 
-public class OtlpOptionsConfig extends OptionsConfig
+public class OtlpOverridesConfig extends OptionsConfig
 {
-    public enum OtlpSignalsConfig
-    {
-        METRICS,
-    }
+    public URI metrics;
 
-    public long interval;
-    public Set<OtlpSignalsConfig> signals;
-    public OtlpEndpointConfig endpoint;
-
-    public OtlpOptionsConfig(
-        long interval,
-        Set<OtlpSignalsConfig> signals,
-        OtlpEndpointConfig endpoint)
+    public OtlpOverridesConfig(
+        URI metrics)
     {
-        this.interval = interval;
-        this.signals = signals;
-        this.endpoint = endpoint;
+        this.metrics = metrics;
     }
 }
